feat(app): workspace UI polish batch — lifecycle/open/logs fixes + persistence, updater button, HTTP-ready
Wave 1 — bug fixes:
- CPU metric emitted a garbage float (f32→f64 serialize repr); cast to f64 before rounding (metrics.rs)
- Open button link list drifted from backend routes; new resolvedTargets() mirrors App::all_routes (local+public groups), popover when >1 host (AppCard)
- Restart/Stop showed during initial Start with no loading; add "Starting…" spinner, keep Stop enabled as Cancel (AppCard)
- Log severity badge pills were ragged and line numbers overflowed the gutter; fixed-width centered pills + digit-aware gutter (LogViewer)
- Git badge was hidden on docker/compose repos; guard on root_dir && !proxy instead of kind==='process' (AppCard)
- extension_shell_run could hang on a git cred prompt and leak child processes; GIT_TERMINAL_PROMPT=0 + stdin null, own process group with group-kill on timeout, concurrent stdout/stderr drain (extension_shell.rs)
- Reorder IPC was fire-and-forget; await + reload-on-error to stop memory/DB divergence (workspace/service slices)
- AppCard re-detected extensions on every workspace switch (shimmer); seed from the store cache, skip re-detect when warm (AppCard)

Wave 2 — small builds:
- Persist sidebar collapse/expand + per-app instances state to localStorage via the ui slice
- Update button in the sidebar footer above the gear (version-labelled), wired to the updater; drop the redundant app name in About
- Log severity rail now spans the whole error/warn block instead of only continuation lines
- app:ready gated on real HTTP readiness (health path, else GET /) instead of the first TCP accept; immediate health refresh on ready

+++ src-tauri/src/metrics.rs
@@ -48,7 +48,7 @@ pub fn spawn_metrics_poller(app: tauri::AppHandle) {
             for app_id in &docker_ids {
                 if let Some((cpu, mem)) = DockerManager::stats(app_id) {
                     let payload = serde_json::json!({
-                        "cpu": (cpu * 10.0).round() / 10.0,
+                        "cpu": ((cpu as f64) * 10.0).round() / 10.0,
                         "mem_mb": (mem as f64 / 1_048_576.0).round() as u64,
                     });
                     app.emit(&format!("app:metrics:{}", app_id), payload).ok();
@@ -71,7 +71,7 @@ pub fn spawn_metrics_poller(app: tauri::AppHandle) {
 
                 if cpu > 0.0 || mem > 0 {
                     let payload = serde_json::json!({
-                        "cpu": (cpu * 10.0).round() / 10.0,  // 1 decimal place
+                        "cpu": ((cpu as f64) * 10.0).round() / 10.0,  // 1 decimal place
                         "mem_mb": (mem as f64 / 1_048_576.0).round() as u64,
                     });
                     app.emit(&format!("app:metrics:{}", app_id), payload).ok();
